requirements.txt
asgiref==3.10.0
diff-match-patch==20241021
Django==5.2.7
django-ckeditor==6.7.3
django-import-export==4.3.12
django-js-asset==3.1.2
pillow==12.0.0
pytube==15.0.0
sqlparse==0.5.3
tablib==3.9.0
tzdata==2025.2
